+++ backend/lib/edgehog/astarte.ex
@@ -27,6 +27,8 @@ defmodule Edgehog.Astarte do
   alias Edgehog.Repo
 
   alias Astarte.Client.AppEngine
+  alias Astarte.Client.APIError
+  alias Astarte.Client.RealmManagement
   alias Edgehog.Astarte.Cluster
   alias Edgehog.Astarte.Device
   alias Edgehog.Astarte.InterfaceVersion
@@ -46,6 +48,11 @@ defmodule Edgehog.Astarte do
     StorageUsage,
     SystemStatus,
     WiFiScanResult
+  }
+
+  alias Edgehog.Astarte.Realm.{
+    Interfaces,
+    Triggers
   }
 
   @ota_request_interface "io.edgehog.devicemanager.OTARequest"
@@ -109,6 +116,9 @@ defmodule Edgehog.Astarte do
                               :astarte_network_interface_module,
                               NetworkInterface
                             )
+
+  @interfaces_module Application.compile_env(:edgehog, :astarte_interfaces_module, Interfaces)
+  @triggers_module Application.compile_env(:edgehog, :astarte_triggers_module, Triggers)
 
   @doc """
   Returns the list of clusters.
@@ -588,6 +598,60 @@ defmodule Edgehog.Astarte do
 
   def send_led_behavior(%AppEngine{} = client, device_id, behavior) do
     @led_behavior_module.post(client, device_id, behavior)
+  end
+
+  def fetch_interface(%RealmManagement{} = client, interface_name, interface_major) do
+    case @interfaces_module.get(client, interface_name, interface_major) do
+      {:ok, %{"data" => interface_map}} ->
+        {:ok, interface_map}
+
+      {:error, %APIError{status: 404}} ->
+        # We just convert a 404 status to :not_found, and return all other errors as-is
+        {:error, :not_found}
+
+      {:error, other} ->
+        {:error, other}
+    end
+  end
+
+  def create_interface(%RealmManagement{} = client, interface_json) do
+    @interfaces_module.create(client, interface_json)
+  end
+
+  def update_interface(
+        %RealmManagement{} = client,
+        interface_name,
+        interface_major,
+        interface_json
+      ) do
+    @interfaces_module.update(client, interface_name, interface_major, interface_json)
+  end
+
+  def fetch_trigger(%RealmManagement{} = client, trigger_name) do
+    case @triggers_module.get(client, trigger_name) do
+      {:ok, %{"data" => trigger_map}} ->
+        {:ok, trigger_map}
+
+      {:error, %APIError{status: 404}} ->
+        # We just convert a 404 status to :not_found, and return all other errors as-is
+        {:error, :not_found}
+
+      # TODO: workaround due to the fact that currently Astarte RM API returns 500 for a
+      # non-existing trigger
+      {:error, %APIError{status: 500}} ->
+        {:error, :not_found}
+
+      {:error, other} ->
+        {:error, other}
+    end
+  end
+
+  def create_trigger(%RealmManagement{} = client, trigger_json) do
+    @triggers_module.create(client, trigger_json)
+  end
+
+  def delete_trigger(%RealmManagement{} = client, trigger_name) do
+    @triggers_module.delete(client, trigger_name)
   end
 
   # get_device_status is already called to fetch other info from the device