+++ backend/test/support/mocks.ex
@@ -74,6 +74,14 @@ Mox.defmock(Edgehog.Astarte.Device.LedBehaviorMock,
   for: Edgehog.Astarte.Device.LedBehavior.Behaviour
 )
 
+Mox.defmock(Edgehog.Astarte.Realm.InterfacesMock,
+  for: Edgehog.Astarte.Realm.Interfaces.Behaviour
+)
+
+Mox.defmock(Edgehog.Astarte.Realm.TriggersMock,
+  for: Edgehog.Astarte.Realm.Triggers.Behaviour
+)
+
 Mox.defmock(Edgehog.Geolocation.GeolocationProviderMock,
   for: Edgehog.Geolocation.GeolocationProvider
 )